Refactor project structure and add background service worker
