+++ config/webpack.config.js
@@ -8,8 +8,8 @@ const PATHS = require('./paths');
 // Merge webpack configuration files
 const config = merge(common, {
   entry: {
-    popup: PATHS.src + '/popup.ts',
-    contentScript: PATHS.src + '/contentScript.js'
+    popup: PATHS.src + '/popup/popup.ts',
+    background: PATHS.src + '/background/index.ts'
   },
 });
