+++ public/popup/popup.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8"/>
     <title>Minha Extensão</title>
-    <script src="popup.js"></script>
-    <link rel="stylesheet" href="popup.css">
+    <script src="../popup.js"></script>
+    <link rel="stylesheet" href="../popup.css">
   </head>
   <body>
     <div id="login-container" style="display:none">
